move state storage configuration functions to a commonly accessible location
