+++ ydb/core/tx/scheme_board/populator_ut.cpp
@@ -1,17 +1,14 @@
 #include "events_schemeshard.h"
 #include "populator.h"
-#include "replica.h"
 #include "ut_helpers.h"
 
 #include <ydb/core/base/statestorage_impl.h>
-#include <ydb/core/base/tablet_types.h>
 #include <ydb/core/scheme/scheme_pathid.h>
 #include <ydb/core/testlib/actors/block_events.h>
 #include <ydb/core/testlib/actors/wait_events.h>
 #include <ydb/core/tx/schemeshard/ut_helpers/helpers.h>
 
 #include <library/cpp/testing/unittest/registar.h>
-#include <ydb/library/actors/core/log.h>
 #include <ydb/library/services/services.pb.h>
 
 #include <util/generic/map.h>
@@ -133,99 +130,8 @@ UNIT_TEST_SUITE_REGISTRATION(TPopulatorTestWithResets);
 
 Y_UNIT_TEST_SUITE(TPopulatorQuorumTest) {
 
-    using TUpdateAck = NSchemeshardEvents::TEvUpdateAck;
-
-    void AddReplicas(TStateStorageInfo::TRingGroup& group, const TVector<TActorId>& replicas) {
-        group.NToSelect = replicas.size();
-        group.Rings.resize(replicas.size());
-        for (size_t i = 0; i < replicas.size(); ++i) {
-            group.Rings[i].Replicas.push_back(replicas[i]);
-        }
-    }
-
-    TIntrusivePtr<TStateStorageInfo> GenerateStateStorageInfo(const TVector<TStateStorageInfo::TRingGroup>& ringGroups) {
-        auto info = MakeIntrusive<TStateStorageInfo>();
-        info->RingGroups = ringGroups;
-        return info;
-    }
-
-    using TStateStorageSetupper = std::function<void(TTestActorRuntime&, ui32)>;
-
     constexpr int ReplicasInRingGroup = 3;
-
-    TStateStorageSetupper CreateCustomStateStorageSetupper(const TVector<TStateStorageInfo::TRingGroup>& ringGroups) {
-        return [ringGroups](TTestActorRuntime& runtime, ui32 nodeIndex) {
-            auto ssInfo = GenerateStateStorageInfo(ringGroups);
-            auto sbInfo = GenerateStateStorageInfo(ringGroups);
-            auto bInfo = GenerateStateStorageInfo(ringGroups);
-
-            for (size_t ringGroup = 0; ringGroup < ringGroups.size(); ++ringGroup) {
-                TVector<TActorId> ssreplicas;
-                for (int i = 0; i < ReplicasInRingGroup; ++i) {
-                    ssreplicas.emplace_back(MakeStateStorageReplicaID(runtime.GetNodeId(0), ReplicasInRingGroup * ringGroup + i));
-                }
-                AddReplicas(ssInfo->RingGroups[ringGroup], ssreplicas);
-
-                TVector<TActorId> sbreplicas;
-                for (int i = 0; i < ReplicasInRingGroup; ++i) {
-                    sbreplicas.emplace_back(MakeSchemeBoardReplicaID(runtime.GetNodeId(0), ReplicasInRingGroup * ringGroup + i));
-                }
-                AddReplicas(sbInfo->RingGroups[ringGroup], sbreplicas);
-
-                TVector<TActorId> breplicas;
-                for (int i = 0; i < ReplicasInRingGroup; ++i) {
-                    breplicas.emplace_back(MakeBoardReplicaID(runtime.GetNodeId(0), ReplicasInRingGroup * ringGroup + i));
-                }
-                AddReplicas(bInfo->RingGroups[ringGroup], breplicas);
-
-                if (nodeIndex == 0) {
-                    for (ui32 i = 0; i < ReplicasInRingGroup; ++i) {
-                        runtime.AddLocalService(ssreplicas[i],
-                            TActorSetupCmd(CreateStateStorageReplica(ssInfo.Get(), ReplicasInRingGroup * ringGroup + i), TMailboxType::Revolving, 0),
-                            nodeIndex
-                        );
-                        runtime.AddLocalService(sbreplicas[i],
-                            TActorSetupCmd(CreateSchemeBoardReplica(sbInfo.Get(), ReplicasInRingGroup * ringGroup + i), TMailboxType::Revolving, 0),
-                            nodeIndex
-                        );
-                        runtime.AddLocalService(breplicas[i],
-                            TActorSetupCmd(CreateStateStorageBoardReplica(bInfo.Get(), ReplicasInRingGroup * ringGroup + i), TMailboxType::Revolving, 0),
-                            nodeIndex
-                        );
-                    }
-                }
-            }
-
-            const TActorId ssproxy = MakeStateStorageProxyID();
-            runtime.AddLocalService(ssproxy,
-                TActorSetupCmd(CreateStateStorageProxy(ssInfo.Get(), bInfo.Get(), sbInfo.Get()), TMailboxType::Revolving, 0),
-                nodeIndex
-            );
-        };
-    }
-
-    TStateStorageSetupper CreateDefaultStateStorageSetupper() {
-        return CreateCustomStateStorageSetupper({ TStateStorageInfo::TRingGroup{} });
-    }
-
-    void SetupRuntime(TTestActorRuntime& runtime, const TStateStorageSetupper& setupStateStorage = CreateDefaultStateStorageSetupper()) {
-        for (ui32 i : xrange(runtime.GetNodeCount())) {
-            setupStateStorage(runtime, i);
-        }
-        runtime.Initialize(TAppPrepare().Unwrap());
-    }
-
-    TIntrusiveConstPtr<TStateStorageInfo> GetStateStorageConfig(TTestActorRuntime& runtime) {
-        const TActorId recipient = MakeStateStorageProxyID();
-        const TActorId edge = runtime.AllocateEdgeActor();
-        runtime.Send(recipient, edge, new TEvStateStorage::TEvListSchemeBoard(false));
-
-        auto result = runtime.GrabEdgeEvent<TEvStateStorage::TEvListSchemeBoardResult>(edge);
-        UNIT_ASSERT(result);
-        UNIT_ASSERT(result->Get()->Info);
-
-        return result->Get()->Info;
-    }
+    using TUpdateAck = NSchemeshardEvents::TEvUpdateAck;
 
     NKikimrScheme::TEvDescribeSchemeResult CreateSamplePathDescription(ui64 owner, const TPathId& pathId, const TString& path) {
         NKikimrScheme::TEvDescribeSchemeResult desc;
@@ -272,14 +178,14 @@ Y_UNIT_TEST_SUITE(TPopulatorQuorumTest) {
 
     void TestPopulatorQuorum(TVector<TStateStorageInfo::TRingGroup>&& ringGroupsConfiguration) {
         TTestBasicRuntime runtime;
-        SetupRuntime(runtime, CreateCustomStateStorageSetupper(ringGroupsConfiguration));
+        SetupMinimalRuntime(runtime, CreateCustomStateStorageSetupper(ringGroupsConfiguration, ReplicasInRingGroup));
 
         runtime.SetLogPriority(NKikimrServices::SCHEME_BOARD_POPULATOR, NLog::PRI_DEBUG);
 
         const TActorId edge = runtime.AllocateEdgeActor();
         constexpr ui64 owner = TTestTxConfig::SchemeShard;
 
-        auto stateStorageInfo = GetStateStorageConfig(runtime);
+        auto stateStorageInfo = GetStateStorageInfo(runtime);
         const auto replicas = stateStorageInfo->SelectAllReplicas();
         UNIT_ASSERT_VALUES_EQUAL_C(
             replicas.size(),
