+++ ydb/core/tx/scheme_board/subscriber_ut.cpp
@@ -18,14 +18,7 @@ namespace NSchemeBoard {
 
 class TSubscriberTest: public NUnitTest::TTestBase {
     TVector<TActorId> ResolveReplicas() {
-        const TActorId proxy = MakeStateStorageProxyID();
-        const TActorId edge = Context->AllocateEdgeActor();
-
-        Context->Send(proxy, edge, new TEvStateStorage::TEvListSchemeBoard(false));
-        auto ev = Context->GrabEdgeEvent<TEvStateStorage::TEvListSchemeBoardResult>(edge);
-
-        Y_ABORT_UNLESS(ev->Get()->Info);
-        auto allReplicas = ev->Get()->Info->SelectAllReplicas();
+        auto allReplicas = GetStateStorageInfo(*Context)->SelectAllReplicas();
         return TVector<TActorId>(allReplicas.begin(), allReplicas.end());
     }
